requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
elastic-transport==8.13.0
elasticsearch==8.13.0
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.23.6
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
